requirements.txt
pillow
customtkinter
gtts
pandas
pygame